LayoutTests/inspector/debugger/setPauseOnMicrotasks-expected.txt

CONSOLE MESSAGE: BREAKPOINT ACTION LOG 1
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 2
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 3
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 4
Tests for Debugger.setPauseOnMicrotasks command.


== Running test suite: Debugger.setPauseOnMicrotasks
-- Running test case: Debugger.setPauseOnMicrotasks.Disabled.queueMicrotask
PASS: Should not have paused.

-- Running test case: Debugger.setPauseOnMicrotasks.Enabled.queueMicrotask
PASS: Should have paused.

-- Running test case: Debugger.setPauseOnMicrotasks.Disabled.Promise.resolve
PASS: Should not have paused.

-- Running test case: Debugger.setPauseOnMicrotasks.Enabled.Promise.resolve
PASS: Should have paused.

-- Running test case: Debugger.setPauseOnMicrotasks.Disabled.Promise.reject
PASS: Should not have paused.

-- Running test case: Debugger.setPauseOnMicrotasks.Enabled.Promise.reject
PASS: Should have paused.

-- Running test case: Debugger.setPauseOnMicrotasks.Options.Condition

Setting condition to 'false'...

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should not pause.

Setting condition to 'true'...

Triggering breakpoint...
PASS: Should pause.

Triggering breakpoint...
PASS: Should pause.

-- Running test case: Debugger.setPauseOnMicrotasks.Options.IgnoreCount

Setting ignoreCount to '2'...

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should pause.

Triggering breakpoint...
PASS: Should pause.

-- Running test case: Debugger.setPauseOnMicrotasks.Options.Action.Log

Adding log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...
Enabling auto-continue...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

-- Running test case: Debugger.setPauseOnMicrotasks.Options.Actions.Evaluate

Adding evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...
Enabling auto-continue...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.
